requirements.txt
fastapi
pydantic
openai
python-dotenv
uvicorn
aiosqlite